Refactor (#51)
* Change include format
+ Sort includes according to Google code style

* Reformat code
+ Mainly order of includes

* Remove unused includes

* Remove unused enums

* Fix incompatible types
+ Types 'tk_param_t' and 'int' are not compatible

* Fix value may not fit into receiver
+ Values of type 'int64_t' may not fit into the receiver type 'int'

* Fix signedness mismatch

* Fix missing default cases in switches

* Fix potential memory leak

* Fix uninitialised stack variable

* Fix potential memory leak
+ Memory leak could happen if the parser rule rule_eol_opt_n results in an error (the stack is not cleaned)

* Remove empty statement

* Fix signedness mismatch
+ part 2

* Reorganize struct to remove int to enum conversion warning

+++ src/scanner.c
@@ -1,9 +1,8 @@
-#include <stdio.h>
-
 #include "scanner.h"
-#include "scanner_states.h"
-#include "hash_map.h"
+
 #include "error.h"
+#include "hash_map.h"
+#include "scanner_states.h"
 
 #define HMAP_BUCKET_COUNT 53
 
@@ -131,7 +130,7 @@ int scanner_scan(FILE *stream, tklist_t token_list) {
 	token_t tk;
 	while ((tk = scanner_next_token(scanner))->type != TK_EOF) {
 		if (tk->type == TK_ERROR) {
-			ret = tk->param.i;
+			ret = (int)tk->param.i;
 			goto SS_ERROR;
 		}
 		if (!tklist_push_back(token_list, tk)) {
